+++ src/_be.d.ts
@@ -3,11 +3,11 @@ export declare const pending_symbol: unique symbol;
  * Returns a function to ensure that an member key is defined on a ctx object,
  * otherwise it creates the value using the _val factory function.
  */
-export declare function _be<O extends unknown = unknown, C extends object = object>(key: string | symbol, _val: (ctx: C, key: (string | symbol), opts?: _be_opts_type) => (void | O)): be_type<O, C>;
-export declare type Be<O> = (ctx: object, opts?: _be_opts_type) => O;
-export interface _be_opts_type {
+export declare function _be<Output extends unknown = unknown, Ctx extends object = object>(key: string | symbol, _val: (ctx: Ctx, key: (string | symbol), opts?: _be_opts_T) => (void | Output)): Be<Output, Ctx>;
+export declare type Be<Output extends unknown = unknown, Ctx extends object = object> = (ctx: Ctx, opts?: _be_opts_T) => Output;
+export declare type B<Output extends unknown = unknown, Ctx extends object = object> = Be<Output, Ctx>;
+export declare type be_T<Output extends unknown = unknown, Ctx extends object = object> = Be<Output, Ctx>;
+export interface _be_opts_T {
     force?: boolean;
 }
-export declare type B<O> = Be<O>;
-export declare type be_type<O extends unknown = unknown, C extends object = object> = (ctx: C, opts?: _be_opts_type) => O;
 export { _be as _b, };